+++ dobby/RELEASES.md
@@ -1,3 +1,18 @@
+# 0.1.5
+
+- Minor api improvements
+- Add model of include paths
+
+# 0.1.4
+
+- Model executable in compilation database entries because they need to be written to disk
+  when serializing.
+
+# 0.1.3
+
+- Expose API to analyze a plain cdb so that caller can be in charge of deserializing json.
+
+# 0.1.2
 # 0.1.1
 
 - CLI wrapper with `dobby validate` and `dobby info` commands.